Split audio.rs into module files
